cleanup folder structure
