+++ src/main/scala/krs/api/ApiModule.scala
@@ -1,0 +1,7 @@
+package krs.api
+
+import krs.domain.{ DomainModule }
+
+trait ApiModule { this: DomainModule =>
+  val userApi: UserApi = new UserApi(userRepository)
+}
