+++ src/main/scala/krs/infrastructure/InfrastructureModule.scala
@@ -1,8 +1,9 @@
 package krs.infrastructure
 
 import krs.domain._
+import krs.api.{ ApiModule }
 
-trait InfrastructureModule { this: DomainModule =>
+trait InfrastructureModule { this: ApiModule with DomainModule =>
 
   override val userRepository: UserRepository = new UserRepositoryFS("./fixtures/users.json")
