+++ src/main/scala/krs/infrastructure/injector.scala
@@ -1,5 +1,6 @@
 package krs.infrastructure
 
 import krs.domain.{ DomainModule }
+import krs.api.{ ApiModule }
 
-class Injector extends DomainModule with InfrastructureModule
+class Injector extends DomainModule with ApiModule with InfrastructureModule
